+++ tiled/_tests/test_writing.py
@@ -670,7 +670,7 @@ def test_write_with_specified_mimetype(tree):
                     ),
                 ],
             )
-            x.write_partition(df, 0)
+            x.write_partition(0, df)
             x.read()
             x.refresh()
             assert x.data_sources()[0].mimetype == mimetype
@@ -724,8 +724,8 @@ def test_append_partition(
         table_to_append = pyarrow.Table.from_pydict(file_to_append)
 
         x = client.create_appendable_table(orig_table.schema, key="x")
-        x.append_partition(orig_table, 0)
-        x.append_partition(table_to_append, 0)
+        x.append_partition(0, orig_table)
+        x.append_partition(0, table_to_append)
         assert_frame_equal(x.read(), pandas.DataFrame(expected_file), check_dtype=False)
 
 
@@ -765,5 +765,41 @@ def test_create_table_with_custom_name(
                 client.create_appendable_table(table.schema, table_name=table_name)
         else:
             x = client.create_appendable_table(table.schema, table_name=table_name)
-            x.append_partition(table, 0)
+            x.append_partition(0, table)
             assert x.read()["column_name"].to_list() == [1, 2, 3]
+
+
+def test_deprecated_argument_order_raises_warning(tree: CatalogContainerAdapter):
+    table = pyarrow.Table.from_arrays([[1, 2, 3]], ["column_name"])
+    with Context.from_app(build_app(tree)) as context:
+        client = from_context(context, include_data_sources=True)
+
+        # Writable table
+        df = pandas.DataFrame({"A": [1, 2, 3], "B": [4, 5, 6]})
+        structure = TableStructure.from_pandas(df)
+
+        for mimetype in [
+            PARQUET_MIMETYPE,
+            "text/csv",
+            APACHE_ARROW_FILE_MIME_TYPE,
+        ]:
+            x = client.new(
+                "table",
+                [
+                    DataSource(
+                        structure_family=StructureFamily.table,
+                        structure=structure,
+                        mimetype=mimetype,
+                    ),
+                ],
+            )
+            with pytest.warns(DeprecationWarning, match=r"The order of arguments"):
+                x.write_partition(df, 0)  # this argument order is deprecated
+            x.refresh()
+            assert x.read() is not None
+
+        # Appendable table
+        x = client.create_appendable_table(table.schema)
+        with pytest.warns(DeprecationWarning, match=r"The order of arguments"):
+            x.append_partition(table, 0)  # this argument order is deprecated
+        assert x.read() is not None